refactor: Authentication model 데이터 레이어로 이동
